Add repair & clean function

+++ clean.sh
@@ -8,16 +8,23 @@ BFS_ENV_DIR=`cat config.json| jq -r '.BFS_ENV_DIR'`
 BFS_STORAGE_DIR=`cat config.json| jq -r '.BFS_STORAGE_DIR'`
 ProjectDir=`cat config.json| jq -r '.ProjectDir'`
 DBName=`cat config.json| jq -r '.DBName'`
-DBRootPassword=`cat config.json| jq -r '.DBRootPassword'`
+DBUser=`cat config.json| jq -r '.DBUser'`
 DBPassword=`cat config.json| jq -r '.DBPassword'`
 
 if [ "$1"x = "all"x ]; then
-	rm -rf ${BFS_ENV_DIR}
+	rm -rf ${BFS_ENV_DIR}/bfs
 	rm -rf ${BFS_STORAGE_DIR}
+	rm -rf ${BFS_ENV_DIR}/bfs-php-extension
 fi
 
 rm -rf ${ProjectDir}
+rm -rf /etc/httpd/conf.d/${serverName}.conf
 
-mysql -uroot -p${DBRootPassword} -s -e "
-drop database  ${DBName} ;
-quit"
+if [ -z "${DBUser}" ] || [ -z "${DBName}" ] || [ -z "${DBPassword}" ];then
+	echo "Please fill in Database parameters in config.json"
+	exit
+else
+	mysql -u${DBUser} -p${DBPassword} -s -e "
+	drop database ${DBName} ;
+	quit"
+fi
